File: file move to datastore module
